refactor: refactor mapper, add PeriodToTimeMapper
